requirements.txt
black~=24.10.0
Jinja2~=3.1.4
openapi_pydantic~=0.5.0
PyYAML~=5.3.1
convert-case~=1.2.3
